% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RODBC.R
\name{odbcGetConnectionTimeout}
\alias{odbcGetConnectionTimeout}
\title{Get the current ODBC connection timeout}
\usage{
odbcGetConnectionTimeout(channel)
}
\arguments{
\item{channel}{connection handle as returned by \code{\link{odbcConnect}} and \code{\link{odbcDriverConnect}}}
}
\value{
Returns the current connection timeout in seconds. 0 (zero) means there is no timeout
}
\description{
Gets the number of seconds to wait for any request on the connection
to complete before returning to the application.
}
\details{
The connection timeout period is used for any request to the data source.
If it expires before the data source responds to the request you will get
an error message.
}
\note{
\strong{Don't confuse the connection timeout with the statement timeout!}
        A statement (SQL code) sent to the data source expires if the data source
        does not return the result set within thestatement timeout period.
        The connection timeout period expires if the statement could not be
        sent to the data source within the connection timeout period.
}
\seealso{
\code{\link{odbcSetConnectionTimeout}}
}

