% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RODBC.R
\name{odbcSetConnectionTimeout}
\alias{odbcSetConnectionTimeout}
\title{Set the ODBC connection timeout}
\usage{
odbcSetConnectionTimeout(channel, timeout = 30)
}
\arguments{
\item{channel}{connection handle as returned by \code{\link{odbcConnect}} and \code{\link{odbcDriverConnect}}}

\item{timeout}{The new timeout value in seconds (0 means "no timeout")}
}
\value{
0 = success,
         1 = success but with an info message,
        -1 = error,
        -2 = invalid handle
        
        Use \code{\link{odbcGetErrMsg}} to get the message).
}
\description{
The connection timeout period is used for any request to the data source.
If it expires before the data source responds to the request you will get
an error message.
}
\note{
Not every ODBC driver supports (changing) the connection timeout.
        For more details see the documentation of your ODBC driver.

\strong{Don't confuse the connection timeout with the statement timeout!}
        A statement (SQL code) sent to the data source expires if the data source
        does not return the result set within thestatement timeout period.
        The connection timeout period expires if the statement could not be
        sent to the data source within the connection timeout period.

If your ODBC driver is configured to use a connection pool the connection timeout
        may be ignored or cause unforeseeable timeouts.
}
\seealso{
\code{\link{odbcGetConnectionTimeout}}
}

